[BUGFIX] Fix endless recusion error in TreeView
We reset the mounts of the tree view to an empty array to prevent
the endless recursion error of TYPO3:

Endless recursion detected: TYPO3 has detected an error in the database.
Please fix it manually (e.g. using phpMyAdmin) and change the UID of
tx_cal_category:0 to a new value.

See http://forge.typo3.org/issues/16150 to get more information about
a possible cause.